FIX rs232 parser
The read_fifo_size command would read a character from the rs232
queue and discard it. Because of this behaviour, every other char
was missing when we were trying to use getchar().

This has been fixed by removing this method. Instead, getchar will
return the number of chars in the queue if the data is valid, else
it will return 0.